AAE-38416 Refactor
